+++ cmn/misc.c
@@ -22,6 +22,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <sys/socket.h>
+#include <arpa/inet.h>
 
 #include "misc.h"
 #include "ucoind.h"
@@ -165,6 +167,41 @@ misc_genesis_t misc_get_genesis(const uint8_t *pGenesisHash)
     return ret;
 }
 
+
+/** JSON-RPC送信
+ *
+ */
+int misc_sendjson(const char *pSend, const char *pAddr, uint16_t Port)
+{
+    int retval = -1;
+    struct sockaddr_in sv_addr;
+
+    int sock = socket(PF_INET, SOCK_STREAM, 0);
+    if (sock < 0) {
+        return retval;
+    }
+    memset(&sv_addr, 0, sizeof(sv_addr));
+    sv_addr.sin_family = AF_INET;
+    sv_addr.sin_addr.s_addr = inet_addr(pAddr);
+    sv_addr.sin_port = htons(Port);
+    retval = connect(sock, (struct sockaddr *)&sv_addr, sizeof(sv_addr));
+    if (retval < 0) {
+        close(sock);
+        return retval;
+    }
+    write(sock, pSend, strlen(pSend));
+
+    //受信を待つとDBの都合でロックしてしまうため、すぐに閉じる
+
+    close(sock);
+
+    return 0;
+}
+
+
+/**************************************************************************
+ * debug functions
+ **************************************************************************/
 
 #ifdef APP_DEBUG_MEM
